requirements.txt
django
django-guardian
zope.dottedname
psycopg2
django-model-utils
